src/main/java/com/golovko/backend/security/UserDetailsServiceImpl.java
package com.golovko.backend.security;

import com.golovko.backend.domain.ApplicationUser;
import com.golovko.backend.repository.ApplicationUserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
public class UserDetailsServiceImpl implements UserDetailsService {

    @Autowired
    private ApplicationUserRepository applicationUserRepository;

    @Transactional(readOnly = true)
    @Override
    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
        ApplicationUser user = applicationUserRepository.findByEmail(email);

        if (user == null) {
            throw new UsernameNotFoundException("User with email" + email + " is not found");
        }

        return new UserDetailsImpl(user);
    }
}
